requirements.txt
asgiref==3.3.4
Django==3.2.4
django-admin-sortable2==1.0
django-ckeditor==6.1.0
django-js-asset==1.2.2
django-modeltranslation==0.17.2
django-resized==0.3.11
django-unused-media==0.2.2
Pillow==8.2.0
pytils==0.3
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
